feat(timeframe): split from and to containers
