+++ galasa-ui/src/components/test-runs/timeframe/DateTimePicker.tsx
@@ -6,7 +6,7 @@
 'use client';
 
 import { useDateTimeFormat } from '@/contexts/DateTimeFormatContext';
-import styles from '@/styles/test-runs/TestRunsPage.module.css';
+import styles from '@/styles/test-runs/timeframe/TimeFrameContent.module.css';
 import { LOCALE_TO_FLATPICKR_FORMAT_MAP } from '@/utils/constants/common';
 import { parseAndValidateTime } from '@/utils/timeOperations';
 import {
