+++ galasa-ui/src/components/test-runs/timeframe/DurationFilter.tsx
@@ -8,7 +8,7 @@
 import { TimeFrameValues } from '@/utils/interfaces';
 import { FormGroup, NumberInput } from '@carbon/react';
 import { useTranslations } from 'next-intl';
-import styles from '@/styles/TimeFrameContent.module.css';
+import styles from '@/styles/test-runs/timeframe/TimeFrameContent.module.css';
 
 const MAX_DAYS = 90;
 const MAX_HOURS = 23;
