+++ galasa-ui/src/components/test-runs/timeframe/TimeFrameFilter.tsx
@@ -6,7 +6,7 @@
 'use client';
 
 import { fromToSelectionEnum } from './TimeFrameContent';
-import styles from '@/styles/test-runs/TestRunsPage.module.css';
+import styles from '@/styles/test-runs/timeframe/TimeFrameContent.module.css';
 import { TimeFrameValues } from '@/utils/interfaces';
 import DateTimePicker from './DateTimePicker';
 import { useTranslations } from 'next-intl';
@@ -23,35 +23,20 @@ export default function TimeFrameFilter({
   disabled?: boolean;
 }) {
   const translations = useTranslations('TimeFrameFilter');
+
+  // Determine which set of values to use based on the component's role
+  const isFromSelection = fromToSelection === fromToSelectionEnum.FromSelectionOptions;
+
   return (
     <div className={styles.timeFrameFilterContainer}>
       <DateTimePicker
-        legend={
-          fromToSelection === fromToSelectionEnum.FromSelectionOptions
-            ? translations('from')
-            : translations('to')
-        }
-        date={values.fromDate}
-        time={values.fromTime}
-        amPm={values.fromAmPm}
-        onDateChange={(date) =>
-          handleValueChange(
-            fromToSelection === fromToSelectionEnum.FromSelectionOptions ? 'fromDate' : 'toDate',
-            date
-          )
-        }
-        onTimeChange={(time) =>
-          handleValueChange(
-            fromToSelection === fromToSelectionEnum.FromSelectionOptions ? 'fromTime' : 'toTime',
-            time
-          )
-        }
-        onAmPmChange={(amPm) =>
-          handleValueChange(
-            fromToSelection === fromToSelectionEnum.FromSelectionOptions ? 'fromAmPm' : 'toAmPm',
-            amPm
-          )
-        }
+        legend={isFromSelection ? translations('from') : translations('to')}
+        date={isFromSelection ? values.fromDate : values.toDate}
+        time={isFromSelection ? values.fromTime : values.toTime}
+        amPm={isFromSelection ? values.fromAmPm : values.toAmPm}
+        onDateChange={(date) => handleValueChange(isFromSelection ? 'fromDate' : 'toDate', date)}
+        onTimeChange={(time) => handleValueChange(isFromSelection ? 'fromTime' : 'toTime', time)}
+        onAmPmChange={(amPm) => handleValueChange(isFromSelection ? 'fromAmPm' : 'toAmPm', amPm)}
         disabled={disabled}
       />
     </div>
